+++ backend/src/app/config-service.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 import { sheets_v4, google } from 'googleapis';
-import { SERVICE_ACCOUNT } from '../env';
+import { MAILER_CONFIG, SERVICE_ACCOUNT } from '../env';
 import { FeedInfo, FeedType, RuleInfo, RuleState, Config, AppList, CustomFields, AppInfo, FeedConfig, SdfElementType, JobInfo } from '../types/config';
 import { FeedData } from '../types/types';
 import { RuleEvaluator } from './rule-engine';
@@ -131,6 +131,7 @@ export default class ConfigService {
           break;
       }
     }
+    config.execution.notificationsEnabled = !!MAILER_CONFIG;
   }
 
   private loadRules(values: any[][], config: Config) {